+++ tests/contrib/autoguide/test_advi.py
@@ -8,8 +8,9 @@ from torch.distributions import constraints
 import pyro
 import pyro.distributions as dist
 import pyro.poutine as poutine
-from pyro.contrib.autoguide import (AutoCallable, AutoDelta, AutoDiagonalNormal, AutoDiscreteParallel, AutoGuideList,
-                                    AutoLowRankMultivariateNormal, AutoMultivariateNormal, AutoIAFNormal)
+from pyro.contrib.autoguide import (AutoCallable, AutoDelta, AutoDiagonalNormal, AutoDiscreteParallel,
+                                    AutoGuideList, AutoLaplace, AutoLowRankMultivariateNormal,
+                                    AutoMultivariateNormal, AutoIAFNormal)
 from pyro.infer import SVI, Trace_ELBO, TraceEnum_ELBO, TraceGraph_ELBO
 from pyro.optim import Adam
 from tests.common import assert_equal
@@ -116,6 +117,7 @@ def auto_guide_callable(model):
 
 @pytest.mark.parametrize("auto_class", [
     AutoDelta,
+    AutoLaplace,
     AutoDiagonalNormal,
     AutoMultivariateNormal,
     AutoLowRankMultivariateNormal,
@@ -137,7 +139,7 @@ def test_median(auto_class, Elbo):
 
     median = guide.median()
     assert_equal(median["x"], torch.tensor(0.0), prec=0.1)
-    if auto_class is AutoDelta:
+    if auto_class is AutoDelta or auto_class is AutoLaplace:
         assert_equal(median["y"], torch.tensor(-1.0).exp(), prec=0.1)
     else:
         assert_equal(median["y"], torch.tensor(1.0), prec=0.1)
@@ -186,6 +188,7 @@ def test_quantiles(auto_class, Elbo):
 
 @pytest.mark.parametrize("continuous_class", [
     AutoDelta,
+    AutoLaplace,
     AutoDiagonalNormal,
     AutoMultivariateNormal,
     AutoLowRankMultivariateNormal,
@@ -216,6 +219,7 @@ def test_discrete_parallel(continuous_class):
 
 @pytest.mark.parametrize("auto_class", [
     AutoDelta,
+    AutoLaplace,
     AutoDiagonalNormal,
     AutoMultivariateNormal,
     AutoLowRankMultivariateNormal,
@@ -235,6 +239,7 @@ def test_guide_list(auto_class):
 
 @pytest.mark.parametrize("auto_class", [
     AutoDelta,
+    AutoLaplace,
     AutoDiagonalNormal,
     AutoMultivariateNormal,
     AutoLowRankMultivariateNormal,
@@ -258,6 +263,7 @@ def test_callable(auto_class):
 
 @pytest.mark.parametrize("auto_class", [
     AutoDelta,
+    AutoLaplace,
     AutoDiagonalNormal,
     AutoMultivariateNormal,
     AutoLowRankMultivariateNormal,
